Ensuring (int) overflow doesn't mess up allocation of, or indexing into, distance matrices.  Replacing int with size_t in: Variables in checkZeroDist, computeMLDist (in phyloanalysis.cpp) Variables in Alignment::printDist (pos). Variables in PhyloTree::correctDist (nsqr, pos, i, k). Variables in PhyloTree::computeDist (num_pairs, pos, seq1). Variables in 4-parameter PhyloTree::computeDist (n, nSquared, i). Variables in PhyloTree::computeObsDist (pos, seq1, seq2). Parameters to Alignment member functions: computeStateFreq, getSeqName, isGapOnlySeq Note, Alignment::checkIndenticalSeq should use hashing. Values returned by Alignment member functions: getNSeq Other places there might have been issues...PhyloTree::initializeAllPartialPars
While I was there I made some other Alignment member functions return size_t
(e.g. getNPattern), changed the parameter to isGapOnlySeq to size_t.
Also: denominator calculated in PhyloTree::computeDist_Experimental
now matches that used in the non-experimental distance computation
(so the JC distances they report are consistent, to about four decimal places).

+++ main/phylotesting.cpp
@@ -1838,7 +1838,7 @@ double doKmeansClustering(Params &params, PhyloSuperTree *in_tree,
         }
     }
     
-    int ssize = in_tree->getAlnNSite();
+    size_t ssize = in_tree->getAlnNSite();
     double score = computeInformationScore(lhsum, dfsum, ssize, params.model_test_criterion);
     cout << "k-means score for " << ncluster << " partitions: " << score << " (LnL: " << lhsum << "  " << "df: " << dfsum << ")" << endl;
 
@@ -1925,7 +1925,7 @@ void testPartitionModel(Params &params, PhyloSuperTree* in_tree, ModelCheckpoint
         if (params.partition_type == BRLEN_SCALE)
             dfsum -= 1;
     }
-	int ssize = in_tree->getAlnNSite();
+	size_t  ssize = in_tree->getAlnNSite();
 	int64_t num_model = 0;
     int64_t total_num_model = in_tree->size();
